requirements.txt
requests
fastapi
uvicorn
python-magic
pydantic
python-dotenv
langchain
openai
psycopg2-binary
pgvector
sqlalchemy
langchain
sentence-transformers
openai
dspy
python-toon
postgresql
postgresql-contrib
postgresql-14-pgvector
psutil
textblob